requirements.txt
certifi>=2019.6.16
chardet>=3.0.4
Django>=2.2.4
django-bootstrap4>=0.0.8
django-cors-headers>=3.0.2
django-oauth>=1.1
django-oauth-toolkit>=1.2.0
djangorestframework>=3.9.4
idna>=2.8
mysql-connector>=2.2.9
mysql-connector-python>=8.0.16
oauthlib>=3.0.1
pytz>=2019.1
requests>=2.22.0
rest-condition>=1.0.3
sqlparse>=0.3.0
urllib3>=1.25.3
